requirements.txt
PyYAML~=6.0
requests~=2.31.0
simplejson~=3.19.1
networkx~=3.2.1
matplotlib~=3.8.2
pyecharts~=2.0.3